replace react-bootstrap with vite
