+++ fontend-app/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import StyleOptionsDropdown from './components/StyleOptionsDropdown.js'
-import SideBar from './components/SideBar.js'
+import StyleOptionsDropdown from './components/StyleOptionsDropdown'
+import SideBar from './components/SideBar'
 
 function App() {
   return (
