+++ fontend-app/src/components/SelectedStyle.jsx
